+++ healthyme-pacientes/src/main/java/studio/devbyjose/healthyme_pacientes/repository/PacienteRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 import java.util.Optional;
 
 public interface PacienteRepository extends JpaRepository<Paciente, Long> {
-    Optional<Paciente> findByIdUsuario(Long idUsuario);
-    boolean existsByIdUsuario(Long idUsuario);
+    Optional<Paciente> findByIdUsuario(Integer idUsuario);
+    boolean existsByIdUsuario(Integer idUsuario);
 
     @Query("SELECT FUNCTION('MONTH', p.fechaCreacion), COUNT(p) FROM Paciente p GROUP BY FUNCTION('MONTH', p.fechaCreacion)")
     List<Object[]> countPacientesPorMes();
